Introduce script to compose data streams
This commit introduces new Python script build-scripts/compose_ds.py
that can compose an SCAP source data stream file from its individual
SCAP components. The aim of it is to replace the `oscap ds sds-compose`
command in our build process. The script is specialized for our use
case, unlike `oscap` it can't perform in a general situation, because it
assumes a certain structure.