Use behaviors to recurse down the directories
Recursing a directory using pattern match works, but this approach
makes OVAL overlook that some paths are symlinks.
So, although the check was not reporting symlinks, it was following
them down.

The OVAL was following symlinks to directories and collecting files in
them, as evidenced by fails in test scenario
dir_symlink_incorrect_file.pass.sh, which was failing.

The way to recurse down a directory and avoid symlinks is by defining a
behavior.