Complete dconf_ini_file_fix macro
It wasn't aligned with the bash remediations because it doesn't contain
the step to add a lock to
/etc/dconf/db/local.d/locks/00-security-settings-lock whereas the bash
remediation adds the lock. An exception is rule
gnome_gdm_disable_automatic_login. Therefore I changed the
dconf_ini_file_fix macro so that it looks similar to fix text in the
dconf_gnome_disable_automount_open rule, use it in every dconf settings
and replace its use in fix in the rule gnome_gdm_disable_automatic_login
with a different fix text.