Add firewalld_sshd_port_enabled to SRG-OS-000096-GPOS-00050
Originally, check and remediation for enabling sshd ports were
a part of rule configure_firewalld_ports, but that rule was meant to
be generic. We moved these checks and remediations to
firewalld_sshd_port_enabled. So we add firewalld_sshd_port_enabled
to this control to keep the original intent.