Downgrade OCP4 builder image to fedora 34
We use Quay triggers to build content on each merged PR in
ComplianceAsCode/content and we use this content for testing.

Recently, we started hitting issues building content where the build
fails because the microdnf thread can't start. This appears to be a
related issue to other things in fedora 35 [0].

  https://bugzilla.redhat.com/show_bug.cgi?id=2009047

This commit updates the build image to use fedora 34, which isn't
susceptible to this issue. At the very least we can get the automation
for building content to work and we can come back and update the build
image when we have a working fedora 35 image.