Introduce conditional prefixes (shared blocks of code) and refactor the way platforms are being loaded
When using the from_text method, the boolean expression is evaluated and if it requires
some new CPE item (because of templated CPE items), these items are created. When
the from_yaml method is used to load the platform, remediations are actually not stored
within the platform because they are tied to CPE items. So rather currently loaded
CPE items are consulted and conditionals are created on the fly.

Load CPE items first because they are used when evaluating platforms.