Update rule to allow special bits in library dirs
Rule file_permissions_library_dirs is about preventing group-writable or
world-writable files in the library dirs. The Suid bits and the stick
bit don't need to be stripped.

The default mode of file_permissions is to allow stricter permissions,
so this change will make the template ignore the special bits and remove
only the 'w' bits from group and others.