Stop SSGTS when a wrong domain name is informed
When calling the SSGTS to use libvirt it is necessary to inform the
domain name. If it is not possible to connect in the domain because
the informed domain name is incorrect, the connection function returns
None. However, other following functions depends on a connected domain
and exceptions are raised. After informing the user, there is no sense
in continue when a domain does not exist. It was included a check for this.