Fix macro fixtext_audit_rules_unsuccessful_file_modification_o_creat
There can be different positions of the  O_CREAT argument
in the syscall parameters in different syscalls, so we can't
hardcode "a2" in the macro text, but we need to allow to specify
the position as a macro paramter. This is similar to how the
audit_rules_unsuccessful_file_modification_o_creat template is
parametrized.