Fix audit file contents templated rules.
The template uses file contents and if defined in the middle of the
file, extra blank lines are added to the check causing the checks to
fail with the sample files and even the expected content defined in the
file. Moving the template section to the end of the file makes it
require only one newline as expected.