Follow symlinks in ensure_gpgcheck_local_packages
/etc/yum.conf is a symlink to /etc/dnf/dnf.conf.
The bash remediation preserves the symlink, it follows the remediation
and edits /etc/dnf/dnf.conf
The Ansible remediation break the symlink, it creates a new static file.

This change makes the Ansible task edit the symlink when it is one.
Note that the ini_file module doesn't have an option to follow symlinks.