Update sysctl bash template
Update sysctl bash template to take into account
config entries that contain the '/' character.

This character is used as command delimiter for sed,
which causes an issue when its present on the text
being matched.

The same update was needed inside the bash_replace_or_append
bash macro for the formated output.

Introduce a test for the bash_replace_or_append bash
macro for the case when the value to check and/or
remediate contains forward slashes.

Also update sysctl bash template to take into
account config files in /usr/lib/sysctl.d/ for
products that also check them on the OVAL content.

Additionally introduce minor update to sysctl OVAL
template comments. Now the comments match the files
the objects reference.

Signed-off-by: Federico Ramirez <federico.r.ramirez@oracle.com>