Include bash macro to remove PAM module options
During the review of display_login_attempts rule, it was noticed a case
where undesired options should be removed while configuring a PAM
module in order to achieve the desired behaviour. It was created
necessary macros to properly treat cases like this while preserving the
standard approach to update PAM files even in systems with authselect.