OCP4: Our CI checks for version.yml not version.yaml
We must haved named one of our versioned test files .yaml and then
copied the mistake over and over. Coupled with bugs in our e2e test
suite, this means that they were never evaluated.

Also adds a newline to the test files as that seems to be expected in
CaC/content.