OCP4: Drop not-arch from file_groupowner_ovs_conf_db_lock
This is a workaround for #9077 that makes the
file_groupowner_ovs_conf_db_lock rule an OR-rule between one specific to
s390x and one that's targetting the other arches.

This ensures the generic file_groupowner_ovs_conf_db_lock rule is filtered
properly when rules are selected to profiles and we don't end up with
the file_groupowner_ovs_conf_db_lock in a platform profile, but only in a
node profile where it belongs.