Update Ansible remediation for no_empty_passwords
The Ansible remediation was simplified to use the new Ansible macros
where authselect is present. For this specific rule, it was more
efficient to keep the replace module to remediate systems without
authselect. The replace is safe to remove "nullok" option at once for
all entries in the two relevant files.