Align bash_ensure_pam_module_line to the Ansible equivalent
During the development of Ansible macros for PAM rules it was noticed
that fixing controls when multiple lines match could lead to very
specific issues which may change the expected behaviour of a PAM module.
The macro was changed to a even more conservative approach by avoiding
changes to existing lines but creating the required line in these cases.