Use a list of values in sysctl template
This patch adds an ability to use a list of values instead of a single
value in the sysctlval parameter of the sysctl template.  This is useful
for situations when we want to create a rule that passes for multiple
different sysctl values.  This commit modifies the OVAL for the runtime
configuration.  The runtime configuration will be allowed to be any of
the values in the list.  There is an OR relation between the values.  In
fact, this is a first step to enable multiple values in the sysctlval
parameter in the sysctl template, because we will also need to check the
static configuration, which is not done in this commit.