Update CIS RHEL8 controlfile to v2.0.0
The current version of CIS RHEL8 Benchmark is v2.0.0. However, the
version used in the project was still v1.0.1. This commit updated all
requirements in alignment to v2.0.0. Some requirements were removed,
others included and many of them just have their IDs updated.
Although some pending requirements were updated to automated status
where it was noticed a new rule was already available, this commit is
not intended to update the profile coverage but simply update the
requirements in alignment to v2.0.0. Profile coverage is out of scope in
this commit.