Add STIG rule mapping between releases
In STIG Benchmarks, the release number of rule IDs will change whenever
there is a change in it. So we cannot directly correlate the old rule
with the new rule between Benchmark releases.

But as the SV number of the rule stays the same, let's create a mapping
from SVs to full Rule IDs that we can find related new rule in the new
Benchmark.