Improve remediation which enable pam_faillock.so
When the pam_faillock.so module is enabled by directly editing PAM
files, the two required lines in the auth group were inserted using the
pam_unix.so line as reference. This approach causes problems when more
authentication modules are present. The remediation was updated to
properly insert the second pam_faillock.so line in the auth group.