Replace platform conditionals in whole remediation code
The XCCDF `<fix>` elements can contain XCCDF `<sub>` elements,
therefore to perform a replacement in the whole remediation code
we need to replace strings not only in the `<fix>` element's text
but also in `<sub>` element`s tail.

Not doing so caused that the platform conditionals that occured
after `<sub>` elements were not replaced. As a result, some
remediations were skipped when using a container back end, which
caused the test runs to fail.