Specify output file names
This change enables us to explicitly specify output file names of the
relabel_ids.py script. Instead of inferring the output file names by the
script the output file names will be specified on command line. This way
the usage and outputs of relabel_ids.py will be more understandable.
This change will be used in the CMake command that calls relabel_ids.py.