Change the command which set the max age for passwords
The bash remediation was using the passwd -x command to set the password
max age. However, the package which provides the passwd command is no
longer included by default in some images. On the other hand, the chage
command is more recommended to change user password expiry information
and is more prone to be included in most of the images.