Print ID of the executed check during scan
I wanted to get the actual CaC ID of the rule and print it, but
as we need to do the scan with --stig-viewer option (to match IDs with
STIG Benchmark), we overwrite the CaC rule ID with STIG rule ID.

The second best thing we can use is the ID of the CaC check.
This prints the ID of the check used when it is available.
When the rule is notapplicable or notchecked nothing was checked, thus
we don't have a check ID to print.

The ideal approach would be to be able to use the ARF results from a CaC
evaluation, but unfortunately the --stig-viewer option only changes the
rule IDs in an XCCDF result file.