Improve comparison of empty rule elements
When a Rule element was empty a new line was being added to the rule
text and being compared. If the rules differend in number of checks, the
number of empty lines would differ and appear in the diff comparison.

This improves handling of the empty elements, by stripping the empty
space and only adding to the rule text when there is something to add.