Update the var_check CheckEnumeration value
The value "all" works fine when the variable is composed by one single
element. However, when multiple directories are set in the
pam_faillock.so settings, these multiple elements in the variable are
treated as a single element in the
object_account_password_selinux_faillock_dir, making it impossible to
assess the dir properties. Changing the value to "at least one", the
elements are treated individually and correctly.