Improve the test which check config existence
It was previously used a test to check the existence of a
textfilecontent54_object. Although the tested object was already used in
prvious test, it was noticed some extra steps during the assessment.
Therefore, the test was changed to test an existing local variable,
resulting in a reduction of about 20 seconds in the overal assessment
of all 6 test scenarios.