modify passing of inherited cpe_platform_names when processing rules and groups
before this commit, adding of new CPE names and putting them to cpe_platform_names infrastrucutre was part of the _add_child function.
This function is common for groups and rules and the behavior is not desired.
We should parse inherited platforms into CPE names on the group level, it does not make much sense on the group level.
So this functionality was moved to the add_rule function.