Add rule auditd_data_retention_admin_space_left_percentage
In RHEL 9 STIG, there is a requirement configuring the admin_space_left
option in /etc/audit/auditd.conf, which is different option from the
space_left option. The RHEL 9 STIG configures both of these options,
so we can't just modify rule auditd_data_retention_space_left_percentage
but we need to keep it and create a new rule for the admin_space_left
option.