bugfix/change: ssg etc: improve and use mkdir_p
Directory creation needs to be TOCTTOU resistant because tests etc can be parallel.

Make mkdir_p return true if creation succeeded. If directory was not created, return false.

In many cases checked existence of a entity (os.path.exists), not a
directory. Error if target is not directory.

On error raise.