Relabel IDs as a part of shorthand creation
This change moves the responsibility of the `relabel_ids.py` to
the `build_shorthand.py`. In other words, OVAL and OCIL IDs will
be changed when the shorthand is created, and the 2 aforementioned
scripts are merged. The benefit is that we will avoid saving the
content to file and load it back, which can make the build faster.