Refactor how the controls are calculated
It was used a static list which wasn't updated in order to consider all
available status. Since now it is possible to get the status list from a
centralized place, the list of calculated status will be always
complete. If the status is not found in the controls lists, it will
still be reported as zero, making more transparent the controls file
state.