Separete functions to manipulate lists and output
There calculate_stats function was doing too much. It was created
separeted functions to manage the relevant lists and also to print the
output for the users. Once this refactoring was concluded, it was
noticed the function was no longer necessary. The data is getting more
and more centralized in each commit, allowing a single point of truth.