Reduce test data input
We will remove some data from the fixture used in tests of
ssg.id_translate.IDTranslator.translate() method, because these
data aren't important for the behavior of the method. They only
add noise to the test, we should strive to form just minimal
fixtures in our units. Their purpose isn't to be a valid OVAL.