Add a test with xmldiff
The xmldiff library compares 2 xml trees and returns a diff in a form of
a list of differences which describe specific changes such as added
element or changed attribute, and we can easily create a list of
expected differences and compare them.  Unfortunately, it works with
lxml trees, so we have to convert our xml.etree.elementree trees to lxml
trees. Also, it requires adding new dependencies to our CI.