Refactor json output function in controleval.py
The function was renamed from print_to_json to print_stats_json to be
aligned to other print functions. It was also refactored to reduce code
and no longer calculate stats, but use the already calculated stats.
This refactoring also standardized the json object keys, so it is
necessary to adjust any eventual automation which uses this json output.