Include small function to order output list
When it is used the option to show the controls list, since the output
list comes from a set, the elements are not ordered. It was included a
function to order the output only in the context of show-controls
option. The intention is to make the output easier to be read by the
users but without breaking the intentional use of set objects.