Remediation applicable to all NetworkManager connections
The remediation was initially consirering to set a firewalld zone only
to active NetworkManager connections. However, it is possible that a
system has more valid connection which are simply not in use at the
moment. These inactive connections can be used at some point and if this
happen, they will also be compliant with an explicit firewalld zone
assigned to them. This way it is indeeded ensured all connections have a
firewalld zone assigned.