Adjust the OCP OVN platform detection for 4.12
For some reason, the strings that are used in both multus.conf and
the networks/cluster resource have changed in OCP 4.12.

multus.conf now contains "ovn-k8s-cni-overlay" and networks/cluster now
contains OVNKubernetes.

This allows us to proceed with checks that are only valid for OVN on OCP4.12,
where OVN is the default anyway.