Improve regex to detect ifcfg files
On RHEL7 and probably other distros which rely on ifcfg files by
default, there is a ifcfg file for the loopback interface, which is out
of the scope in this rule and should be ignored. This commit also
improved the wording in a OVAL comment to make it more clear.