Align test scenario scripts with banner_etc_issue rule
The scripts were aligned to the equivalent in the banner_etc_issue rule.
Also, during the tests it was noticed that test scenario scripts which
were explicitly setting the stig or ospp profile on their headers were
failining to remediate. It was happening because the banner_etc_issue_net
rule was not used in any STIG or OSPP profiles. In this context, I
opted to remove the test scenario scripts instead of changing the
profiles. The remaining test scripts already cover the rule quality
tests for the applicable profiles.