allow passing of parameters when creating the platform
Now, when creating a platform, you can specify for example
package[ntp]
This will search for CPE item called "package", ignoring the part in brackets.
The item MUST have the "name" attribute parametrized, so that it can get a new name.
After it finds it, it will load it and create a new item with content where {arg} is replaced by the parameter given to the platform.