Update instructions for TLS cipher suite algorithms
The instructions for checking the TLS cipher suites piped the output
into grep, which didn't actually output the kubelet algorithms.

This commit updates the instructions to output the actual values used by
each kubelet so it's easier for users to compare the values across a set
of kubelets.

Fixes OCPBUGS-3017