Rename method _is_parametrized to reference_is_parametrized
We will remove the leading underscore because we plan to not use it as a
private method in the future. Also, we will rename the method because it
might be misleading that we use the method to parametrize the `Symbol`
class but we are using that to evaluate strings that represent platform
units.  Moreover, the `ref` variable name wasn't complete so it didn't
give a hint about the expected parameter so we will rename it as well.