Documentation about Fedora Packages
There are plenty of good documentation about Fedora processes. However,
the process to update scap-security-guide package happen each two
months and should be conducted by different engineers. Therefore, if a
person is not so used to the Fedora Package process or did it long time
ago, reading the complete documentation to find the proper commands
could take some time. This documentation intends to save time and
define a standard process applicable for scap-security-guide in Fedora.