Remove extra condition in Ansible task
The Ansible task to apply changes related to the custom profile creation
were based on three criteria. One criterion was if the task to restore
the authselect features was not skipped. The respective task to restore
authselect features should be usually executed in most cases because
its conditions are expected to be valid at this point. However, the
task goes through a loop based on the result of another task. So, the
list is expected be defined but in some cases it can be empty. In cases
when the list is empty, it was noticed the task is skipped, even if
their criteria conditions are valid. This behavior affects some of the
following tasks in the Playbook by making them to also be skipped and
resulting in changes not applied.