Drop CIS reference in no_legacy_plus_entries_etc_group rule for SLE15
The reference to the CIS SLE benchmark is not correct. As far as I could track in history those references were added in PR #5807,
where the author stated that he has no access to enterprise linux benchmarks and used the distribution independant benchmark,
in which context the references are more or less valid. BTW the rhel7 reference also looks incorrect to me,
but if did not want to change it without feedback from RedHat's team member.