change meaning of function build_lang_for_templatable
originally, the function did also the file writing
but I need to split it so that I can better influence what to do with the result of templating.
In case of platforms and CPE items, I do not want the filled template to be written into a file,
I want to incorporate it into the CPE item / template and save them to disk as needed.

This commit changes tthe function and adapt templating of rules to the new API.
Templating of platforms and CPEs will be handled in separate commit.