rewrite bash remediation
the remediation is on par with ansible. It tries to fix already entered values or add missing values before dumping new omfwd directive.
This is an attempt to preserve user configuration.

Note that here sed is used to read whole file and search for patterns. The magic behind it is explained here:
https://unix.stackexchange.com/a/669938