Move CPE OVALs
We will move the OVALs that are used as applicability checks to
a separate directory `shared/applicability/oval/` to
distinguish them from other shared OVALs and separate them.

This will prevent getting these OVALs to `ssg-product-oval.xml`
and therefore prevent duplication of these OVALs in the built
data stream.

However, there is a problem that we have some OVALs that are used both
as CPE OVAL applicability checks and as extending definitions for some
OVAL definitions in our rules (for example, they are used in a platform
definition and in the rule `installed_os_is_FIPS_certified` at the same
time). These can't be moved, they only could be copied, but that would
mean duplicate code in the git repository. These cases need to be
addressed individually case-by-case. For this moment, we will keep the
directory `shared/checks/OVAL` as one of sources of the
cpe-oval-unlinked.xml. But the future goal is to not use OVALs from
`shared/checks/OVAL` in `cpe-oval-unlinked.xml` at all.