Introduce the release_helper.py script
During the preparation for the release 0.1.65, the release process was
reviewed and updated in order to make the process more efficient and
easier to be automated. During the release, all steps from the process
were validated and, in parallel, I started this script to help in each
Github task related to the process. The script was now polished and
should be ready to be shared with the community. The intention is to
make repetitive tasks more efficient and remove unnecessary human
efforts, like calculating dates, manually collecting data, etc.