Remove command path discovery with which
While the which command is common in most of the systems, it might not
be installed out-of-box in some environments. On the other hand, chown,
chgrp and chmod are more likely to be present. The bash remediation
already uses some commands without the absolut path, so it was also
removed for the three commands previously mentioned.