Remove check of /var/log/dmesg from OVAL
Unlike the output of the dmesg command, the corresponding file may be
binary, and therefore unsuitable to be examined by the textfilecontent
OVAL test, as that one assumes that the file is encoded in UTF-8.

Moreover, the removed checking regex was quite fragile, which was
probably caused by the NX feature having different name on different
platforms and the corresponding log entry not being designed as
machine-readable.