Remove dependency on real-world RPM packages
The test depended on presence of rpm and having the bash RPM package
installed. Not all systems where we run our tests use rpm as a package
manager. Also, the systems differ in installed bash version and that
would be updated in future. So we can't rely on specific version and
that limited us so that we couldn't test the macro in detail.  Instead
of using a real rpm command, we will use a mock rpm command and that
will just return strings instead of the real query.