controls: Support @-notation for several fields
In order to reuse the GPOS controls between RHEL and RHCOS, we support
the @-notation for control files so that attributes in the control files
can say, e.g. status but also status@rhcos4.

The overriden attribute should always be explicit, meaning that the code
always falls back to the "plain" attribute and only then returns a default.