srg export: Filter rules by product
Not all rules in the `rules` attribute should be included in the
resulting export, but only those that pertain to the product we are
generating.

In order to correctly handle the cases where no rules would end up being
selected and to fall back to the `no_selections_row`, we filter the
contents of `control.selections`. Setting of the attribute is a bit odd,
because the `selections` setter does not really work as a setter, but it
mostly appends. The setter is extended with a special case for setting
empty list which empties all the attributes and then the selection (now
empty) is appended to with the filtered data.

Note that this filtering only works based on `products`, not e.g.
`prodtype`.