Change custom zones check in reaction to OpenSCAP change
If any default zone is modified by the administrator, the respective
zone file is placed in /etc/firewalld/zones dir in order to override
the default zone settings. The same directory is applicable for new
zones created by the administrator. Therefore, all files in this
directory should also allow SSH. This test was updated in a reaction
to https://github.com/OpenSCAP/openscap/issues/1923, which changed the
behavior of xmlfilecontent probe in OpenSCAP 1.3.7.