Introduce variable values with two values for RHEL
In RHEL8.8 and RHEL9.2 it was introduced a new authselect feature to
enable pam_pwhistory module. Previously, it was enabled by creating a
custom authselect profile. Historically, the control flag used to enable
pam_pwhistory.so was "required". However, the authselect feature uses
"requisite". These two control flags are correct and the differences
relies on how the module result will be reported to the user, without
affecting the module action. Therefore, the assessment should also pass
if required is used.