Update test scenarios for accounts_password_last_change_is_in_past
Some CI tests use containers, and those container might not have
commands like passwd. One approach would be to install the package in
the test scenarios, but the name of the package that brings up the
passwd command differs among distros. In favor of simplicity, entries
were manually created in the /etc/shadow file. This is sufficient for
the context of this rule.